requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.3
huggingface-hub==0.32.6
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.2.2
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
